requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
Flask==2.1.2
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Werkzeug==2.1.2
whitenoise==6.2.0
aniso8601==9.0.1
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
Flask==2.1.2
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Werkzeug==2.1.2
whitenoise==6.2.0
